Reshuffle functions and helpers
