+++ vizro-core/examples/_chart-gallery/utils/custom_charts_comp.py
@@ -1,10 +1,13 @@
 """Contains custom components and charts used inside the dashboard."""
 
-from typing import Literal
+from typing import Literal, List
 
 import dash_bootstrap_components as dbc
+import pandas as pd
+import plotly.graph_objects as go
 import vizro.models as vm
 from dash import dcc, html
+from vizro.models.types import capture
 
 try:
     from pydantic.v1 import Field
@@ -59,3 +62,50 @@ class FlexContainer(vm.Container):
         return html.Div(
             id=self.id, children=[component.build() for component in self.components], className="flex-container"
         )
+
+
+@capture("graph")
+def butterfly(data_frame: pd.DataFrame, x1: str, x2: str, y: str):
+    fig = go.Figure()
+    fig.add_trace(
+        go.Bar(
+            x=-data_frame[x1].values,
+            y=data_frame[y],
+            orientation="h",
+            name=x1,
+        )
+    )
+    fig.add_trace(
+        go.Bar(
+            x=data_frame[x2],
+            y=data_frame[y],
+            orientation="h",
+            name=x2,
+        )
+    )
+    fig.update_layout(barmode="relative")
+    return fig
+
+
+@capture("graph")
+def sankey(data_frame: pd.DataFrame, source: str, target: str, value: str, labels: List[str]):
+    fig = go.Figure(
+        data=[
+            go.Sankey(
+                node=dict(
+                    pad=16,
+                    thickness=16,
+                    label=labels,
+                ),
+                link=dict(
+                    source=data_frame[source],
+                    target=data_frame[target],
+                    value=data_frame[value],
+                    label=labels,
+                    color="rgba(205, 209, 228, 0.4)",
+                ),
+            )
+        ]
+    )
+    fig.update_layout(barmode="relative")
+    return fig
